Revert "Fix ppc64le backend and 64K page size support (#2828)"
This reverts commit 765bcc2b1d87050a4ea366bf22e1db075ad5660b.

+++ include/support.h
@@ -52,8 +52,7 @@
 #endif
 
 #ifdef PAGESIZE
-// Some platforms like ppc64 have page sizes of 64K, so uint16_t isn't enough.
-constexpr uint32_t host_pagesize = { PAGESIZE };
+constexpr uint16_t host_pagesize = { PAGESIZE };
 #else
 constexpr uint16_t host_pagesize = 4096;
 #endif
